Memory Mapped Ring Buffers (#6)
* Introduce tests

* enqueue looking good on first pass

* read accessors

* some refactoring

* test confirmed

* maybe?

* tests pass

* progress

* 2 more

* testing

* version bump

+++ Intrinio.Collections/RingBuffers/DelayDynamicBlockDropOldestRingBuffer.cs
@@ -16,13 +16,13 @@ public class DelayDynamicBlockDropOldestRingBuffer: IDynamicBlockRingBuffer
     private readonly long[] _enqueueTimes;
     private readonly long _delayMilliseconds;
     private readonly System.Diagnostics.Stopwatch _stopwatch;
-    private uint _blockNextReadIndex;
-    private uint _blockNextWriteIndex;
+    private ulong _blockNextReadIndex;
+    private ulong _blockNextWriteIndex;
     private readonly object _readLock;
     private readonly object _writeLock;
     private ulong _count;
     private readonly uint _blockSize;
-    private readonly uint _blockCapacity;
+    private readonly ulong _blockCapacity;
     private ulong _dropCount;
     
     private ulong _processed;
@@ -30,7 +30,7 @@ public class DelayDynamicBlockDropOldestRingBuffer: IDynamicBlockRingBuffer
     
     public ulong Count { get { return Interlocked.Read(ref _count); } }
     public uint BlockSize { get { return _blockSize; } }
-    public uint BlockCapacity { get { return _blockCapacity; } }
+    public ulong BlockCapacity { get { return _blockCapacity; } }
     public ulong DropCount { get { return Interlocked.Read(ref _dropCount); } }
 
     public bool IsEmpty
@@ -56,10 +56,10 @@ public class DelayDynamicBlockDropOldestRingBuffer: IDynamicBlockRingBuffer
     /// A read thread-safe, write not thread-safe implementation of the <see cref="IDynamicBlockRingBuffer"/> (single producer and multiple consumer).  Full behavior: the block trying to be enqueued will be dropped. Provides support for dealing with blocks of varying size less than or equal to block size. 
     /// </summary>
     /// <param name="delayMilliseconds">The number of milliseconds to delay blocks from being dequeued.</param>
-    /// <param name="blockSize">The fixed size of each byte block. Internally, the first sizeof(UInt32) of each block is reserved for tracking the used size of each block. /></param>
+    /// <param name="blockSize">The fixed size of each byte block.  /></param>
     /// <param name="blockCapacity">The fixed capacity of block count.</param>
     /// <param name="stopwatch">The stopwatch to use for comparison of elapsed milliseconds at time of dequeue to the elapsed milliseconds at the time a partition block was enqueued.</param>
-    public DelayDynamicBlockDropOldestRingBuffer(uint delayMilliseconds, uint blockSize, uint blockCapacity, System.Diagnostics.Stopwatch? stopwatch = default)
+    public DelayDynamicBlockDropOldestRingBuffer(uint delayMilliseconds, uint blockSize, ulong blockCapacity, System.Diagnostics.Stopwatch? stopwatch = default)
     {
         _blockSize = blockSize;
         _blockCapacity = blockCapacity;
