+++ Intrinio.Collections/RingBuffers/RingBuffer.cs
@@ -11,20 +11,20 @@ public class RingBuffer : IRingBuffer
 {
     #region Data Members
     private readonly byte[] _data;
-    private uint _blockNextReadIndex;
-    private uint _blockNextWriteIndex;
+    private ulong _blockNextReadIndex;
+    private ulong _blockNextWriteIndex;
     private readonly object _readLock;
     private readonly object _writeLock;
     private ulong _count;
     private readonly uint _blockSize;
-    private readonly uint _blockCapacity;
+    private readonly ulong _blockCapacity;
     private ulong _dropCount;
     private ulong _processed;
     public ulong ProcessedCount { get { return Interlocked.Read(ref _processed); } }
     
     public ulong Count { get { return Interlocked.Read(ref _count); } }
     public uint BlockSize { get { return _blockSize; } }
-    public uint BlockCapacity { get { return _blockCapacity; } }
+    public ulong BlockCapacity { get { return _blockCapacity; } }
     public ulong DropCount { get { return Interlocked.Read(ref _dropCount); } }
 
     public bool IsEmpty
@@ -51,7 +51,7 @@ public class RingBuffer : IRingBuffer
     /// </summary>
     /// <param name="blockSize">The fixed size of each byte block.</param>
     /// <param name="blockCapacity">The fixed capacity of block count.</param>
-    public RingBuffer(uint blockSize, uint blockCapacity)
+    public RingBuffer(uint blockSize, ulong blockCapacity)
     {
         this._blockSize = blockSize;
         this._blockCapacity = blockCapacity;
